test: move mise package version extraction to mise step builder tests
