+++ core/mise/mise_test.go
@@ -2,39 +2,11 @@ package mise
 
 import (
 	"os"
-	"path/filepath"
 	"strings"
 	"testing"
 
-	"github.com/railwayapp/railpack/core/logger"
 	"github.com/stretchr/testify/require"
 )
-
-type testContext struct {
-	appSource string
-	logger    *logger.Logger
-}
-
-func (t *testContext) GetAppSource() string {
-	return t.appSource
-}
-
-func (t *testContext) GetLogger() *logger.Logger {
-	return t.logger
-}
-
-func CreateTestContext(t *testing.T, appDir string) MiseAppContext {
-	absAppDir, err := filepath.Abs(appDir)
-	if err != nil {
-		t.Fatalf("failed to get absolute path: %v", err)
-	}
-
-	logger := logger.NewLogger()
-	return &testContext{
-		appSource: absAppDir,
-		logger:    logger,
-	}
-}
 
 func TestMistGetLatestVersion(t *testing.T) {
 	tempDir, err := os.MkdirTemp("", "mise-test")
@@ -151,75 +123,4 @@ func TestMiseGetAllVersions(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestGetPackageVersions(t *testing.T) {
-	tempDir, err := os.MkdirTemp("", "mise-test")
-	if err != nil {
-		t.Fatalf("failed to create temp dir: %v", err)
-	}
-	defer os.RemoveAll(tempDir)
-
-	mise, err := New(tempDir)
-	if err != nil {
-		t.Fatalf("failed to create mise: %v", err)
-	}
-
-	ctx := CreateTestContext(t, "../../examples/python-uv-tool-versions")
-
-	packages, err := mise.GetPackageVersions(ctx)
-	require.NoError(t, err)
-
-	// Expected packages from the example
-	expected := map[string]struct{}{
-		"python": {},
-		"uv":     {},
-	}
-
-	// Ensure ONLY expected packages are present
-	require.Len(t, packages, len(expected), "unexpected number of packages returned")
-	for name := range packages {
-		_, ok := expected[name]
-		require.True(t, ok, "unexpected package found: %s", name)
-	}
-
-	// The python-uv-tool-versions example should have python and uv defined
-	require.Contains(t, packages, "python")
-	require.Contains(t, packages, "uv")
-
-	// Verify versions are not empty
-	require.NotEmpty(t, packages["python"].Version)
-	require.NotEmpty(t, packages["uv"].Version)
-
-	// Verify python version starts with "3.9" (as defined in .tool-versions)
-	require.True(t, strings.HasPrefix(packages["python"].Version, "3.9"))
-
-	// Verify uv version starts with "0.7" (as defined in .tool-versions)
-	require.True(t, strings.HasPrefix(packages["uv"].Version, "0.7"))
-
-	// Verify source types are set
-	require.NotEmpty(t, packages["python"].Source)
-	require.NotEmpty(t, packages["uv"].Source)
-}
-
-func TestGetPackageVersionsWithNoToolVersions(t *testing.T) {
-	tempDir, err := os.MkdirTemp("", "mise-test")
-	if err != nil {
-		t.Fatalf("failed to create temp dir: %v", err)
-	}
-	defer os.RemoveAll(tempDir)
-
-	mise, err := New(tempDir)
-	if err != nil {
-		t.Fatalf("failed to create mise: %v", err)
-	}
-
-	// Use an example directory that doesn't have a .tool-versions file
-	ctx := CreateTestContext(t, "../../examples/node-tanstack-start")
-
-	packages, err := mise.GetPackageVersions(ctx)
-	require.NoError(t, err)
-
-	// Should return empty map for directory with no .tool-versions
-	require.Empty(t, packages)
 }
